Keep the all examples inn one file, allow users to run them with preferred vector store. Record the actual run rather then mock for E2E test
