+++ tests/test_rag_adapter/test_rag_end_to_end.py
@@ -244,6 +244,7 @@ def test_rag_end_to_end_optimization(sample_ai_ml_dataset, mock_chromadb_store):
         reflection_lm=simple_reflection_lm,
         display_progress_bar=True,
     )
+    
 
     # 3. Assertions: Verify the optimization completed successfully
     assert gepa_result is not None
@@ -272,6 +273,129 @@ def test_rag_end_to_end_optimization(sample_ai_ml_dataset, mock_chromadb_store):
     # Verify GEPA completed metric calls and evaluations
     assert gepa_result.total_metric_calls > 0
     assert gepa_result.num_full_val_evals > 0
+    
+    # ==== Enhanced GEPA workflow validation (addresses PR feedback) ====
+    
+    # 1. Verify expected GEPA workflow structure - these values encode current behavior
+    assert gepa_result.total_metric_calls == 7  # Actual observed value from test run
+    assert gepa_result.num_full_val_evals == 1  # Only full validation evals, not all metric calls
+    assert gepa_result.best_idx == 0  # First (base) program should be best in this deterministic test
+    
+    # 2. Verify score arrays have expected structure
+    assert len(gepa_result.val_aggregate_scores) == 1  # Only one validation score recorded
+    assert len(gepa_result.val_subscores) == 1  # Val subscores for each candidate
+    assert len(gepa_result.val_subscores[0]) == 1  # One score per validation instance
+    assert all(isinstance(score, (int, float)) for score in gepa_result.val_aggregate_scores)
+    assert all(isinstance(subscores, list) for subscores in gepa_result.val_subscores)
+    assert all(0 <= score <= 1 for score in gepa_result.val_aggregate_scores)  # Scores should be normalized
+    
+    # 3. Verify base program evaluation (iteration 0) happened
+    base_val_score = gepa_result.val_aggregate_scores[0]  
+    base_val_subscore = gepa_result.val_subscores[0][0]  # First candidate, first val instance
+    assert isinstance(base_val_score, (int, float))
+    assert isinstance(base_val_subscore, (int, float))
+    assert base_val_score > 0  # Should have meaningful score from mock LLM
+    assert base_val_score == base_val_subscore  # Should match since only one val instance
+    
+    # 4. Verify GEPA attempted optimization iterations (metric calls include subsampling)
+    assert gepa_result.total_metric_calls > gepa_result.num_full_val_evals  # Should have done subsampling
+    # With 7 total metric calls but only 1 full val eval, GEPA did internal optimization work
+    
+    # 5. Verify best candidate selection logic worked
+    best_val_score = gepa_result.val_aggregate_scores[gepa_result.best_idx]
+    assert best_val_score == max(gepa_result.val_aggregate_scores)  # Best idx should point to highest score
+    
+    # 6. Encode expected optimized prompt structure based on reflection_lm output
+    # Our simple_reflection_lm returns a specific JSON structure - verify GEPA used it correctly
+    expected_optimized_prompt_contains = [
+        "comprehensive",  # From our reflection LLM output
+        "accurate", 
+        "Context:",  # Should preserve template variables
+        "{context}",  # Template variable should be maintained
+        "{query}"    # Template variable should be maintained  
+    ]
+    
+    # The reflection may or may not be accepted, but if a new prompt was generated, it should have these characteristics
+    final_prompt = gepa_result.best_candidate["answer_generation"]
+    if final_prompt != seed_candidate["answer_generation"]:
+        # If optimization occurred, verify the expected structure
+        for expected_text in expected_optimized_prompt_contains[:2]:  # Check first 2 elements
+            assert expected_text.lower() in final_prompt.lower(), f"Expected '{expected_text}' in optimized prompt"
+        
+        # Verify template variables are preserved
+        assert "{context}" in final_prompt, "Template variable {context} should be preserved"
+        assert "{query}" in final_prompt, "Template variable {query} should be preserved"
+    
+    # 7. Verify reproducibility - with same mock functions, results should be deterministic
+    # Run a second optimization with identical setup
+    adapter2 = TestRAGAdapter(vector_store=mock_chromadb_store, rag_config=rag_config)
+    gepa_result2 = gepa.optimize(
+        seed_candidate=seed_candidate,
+        trainset=trainset,
+        valset=valset,
+        adapter=adapter2,
+        max_metric_calls=5,
+        reflection_lm=simple_reflection_lm,
+        display_progress_bar=False,  # Disable progress bar for cleaner test output
+    )
+    
+    # Results should be identical due to deterministic mocks
+    assert gepa_result2.total_metric_calls == gepa_result.total_metric_calls
+    assert gepa_result2.num_full_val_evals == gepa_result.num_full_val_evals
+    assert gepa_result2.best_idx == gepa_result.best_idx
+    assert gepa_result2.val_aggregate_scores == gepa_result.val_aggregate_scores
+    assert gepa_result2.best_candidate == gepa_result.best_candidate
+    
+    # 8. Record and assert exact expected results (regression test for GEPA workflow)
+    # These values are captured from actual runs and should remain stable unless GEPA logic changes
+    EXPECTED_EXACT_RESULTS = {
+        "total_metric_calls": 7,  # Updated from actual test run
+        "num_full_val_evals": 1,   # Updated from actual test run
+        "best_idx": 0,  # Base program should be best with our deterministic setup
+        # Exact scores depend on the mock LLM responses and evaluation logic
+        # These values were captured from an actual test run  
+        "expected_val_score": 0.6637837837837838,  # Exact score from deterministic run
+        # The final prompt should match exactly - recorded from actual run
+        "seed_prompt": "Answer the question '{query}' using the provided context: {context}",
+        "expected_final_prompt": "Answer the question '{query}' using the provided context: {context}",  # No change in this deterministic case
+        "expected_prompt_changed": False,  # Reflection didn't improve score, so prompt wasn't changed
+        # Record what the reflection LLM proposed (even though it wasn't accepted)
+        "proposed_optimized_prompt": "Based on the provided context, give a comprehensive and accurate answer to the question '{query}'. Structure your response clearly, include key definitions and explanations, and ensure your answer directly addresses all aspects of the question. Context: {context}",
+    }
+    
+    # Assert exact workflow values match expected
+    assert gepa_result.total_metric_calls == EXPECTED_EXACT_RESULTS["total_metric_calls"]
+    assert gepa_result.num_full_val_evals == EXPECTED_EXACT_RESULTS["num_full_val_evals"] 
+    assert gepa_result.best_idx == EXPECTED_EXACT_RESULTS["best_idx"]
+    
+    # Assert exact score from recorded run
+    expected_val_score = EXPECTED_EXACT_RESULTS["expected_val_score"]
+    assert base_val_score == expected_val_score, f"Base validation score {base_val_score} != expected {expected_val_score}"
+    
+    # 9. Assert exact final prompt (the key assertion requested in PR feedback)
+    # Record the ACTUAL optimized prompt from the deterministic run
+    expected_final_prompt = EXPECTED_EXACT_RESULTS["expected_final_prompt"]
+    actual_final_prompt = gepa_result.best_candidate["answer_generation"]
+    prompt_changed = (seed_candidate["answer_generation"] != actual_final_prompt)
+    
+    # Assert the exact prompt matches what we recorded
+    assert actual_final_prompt == expected_final_prompt, \
+        f"Final prompt doesn't match recorded result:\nActual: {repr(actual_final_prompt)}\nExpected: {repr(expected_final_prompt)}"
+    
+    # Assert whether prompt changed matches expected behavior
+    expected_changed = EXPECTED_EXACT_RESULTS["expected_prompt_changed"]
+    assert prompt_changed == expected_changed, \
+        f"Prompt change behavior {prompt_changed} != expected {expected_changed}"
+    
+    # Document what GEPA's reflection LLM proposed (for reference, even if not accepted)
+    proposed_prompt = EXPECTED_EXACT_RESULTS["proposed_optimized_prompt"]
+    print(f"\n💡 GEPA Workflow Record:")
+    print(f"   Seed prompt: {repr(seed_candidate['answer_generation'])}")
+    print(f"   Proposed by reflection: {repr(proposed_prompt)}")  
+    print(f"   Final prompt: {repr(actual_final_prompt)}")
+    print(f"   Prompt was changed: {prompt_changed}")
+    print(f"   Why unchanged: Reflection proposal didn't improve validation score")
+
 
 
 def test_rag_adapter_basic_functionality(mock_chromadb_store):
